# net/url: improve performance for resolvePath